// block-performance-1.0.0/src/library/cloudman/app-build/type.ts
export interface Input {
    name: string;
    version: string;
    library_list: string[];
    npm_script: { dev: string; start: string; stop: string };
    customize_file_list: { name: string; json_data?: { [key: string]: any } }[];
}

export interface OutputOk {
    target_dir: string;
    target_tar_zip_file: string;
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}